+++ oap-server/server-core/src/test/java/org/apache/skywalking/oap/server/core/management/ui/template/UITemplateTest.java
@@ -18,8 +18,6 @@
 
 package org.apache.skywalking.oap.server.core.management.ui.template;
 
-import org.apache.skywalking.oap.server.core.query.enumeration.TemplateType;
-import org.apache.skywalking.oap.server.library.util.BooleanUtils;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -27,11 +25,8 @@ public class UITemplateTest {
     @Test
     public void testSerialization() {
         UITemplate uiTemplate = new UITemplate();
-        uiTemplate.setName("name");
+        uiTemplate.setTemplateId("id");
         uiTemplate.setConfiguration("configuration");
-        uiTemplate.setType(TemplateType.DASHBOARD.name());
-        uiTemplate.setActivated(BooleanUtils.TRUE);
-        uiTemplate.setDisabled(BooleanUtils.FALSE);
 
         final UITemplate.Builder builder = new UITemplate.Builder();
         final UITemplate uiTemplate2 = builder.storage2Entity(builder.entity2Storage(uiTemplate));
@@ -39,10 +34,7 @@ public class UITemplateTest {
         Assert.assertEquals(uiTemplate, uiTemplate2);
 
         uiTemplate2.setConfiguration("configuration2");
-        uiTemplate.setType(TemplateType.TOPOLOGY_ENDPOINT.name());
-        uiTemplate.setActivated(BooleanUtils.FALSE);
-        uiTemplate.setDisabled(BooleanUtils.TRUE);
-        // Equals method is only for `name` field.
+        // Equals method is only for `templateId` field.
         Assert.assertEquals(uiTemplate, uiTemplate2);
     }
 }